sergi to main

+++ api/laravel/app/Http/Controllers/PropietatsController.php
@@ -50,12 +50,13 @@ class PropietatsController extends Controller
             }
 
             $propietat->es_principal = $request->input('es_principal');
+
             $propietat->preu = $request->input('preu');
             $propietat->stock = $request->input('stock');
 
             // si el stock es mayor a 0 enviamos correo
             /* if ($propietat->stock > 0) {
-                $correos = App\Http\Controllers\CorreuNoStock::where('id_propietats', $request->input('id_propietats'))->get();
+                $correos = CorreuNo::where('id_propietats', $request->input('id_propietats'))->get();
                 foreach ($correos as $c) {
                     $email = $c->email;
                     $correo = new PCEconomic();
@@ -72,34 +73,51 @@ class PropietatsController extends Controller
                 return response()->json(['message' =>
                 'El preu i el stock no poden estar buits'], 400);
             }
-            $var = array_keys($request->input('propietats_valors'));
-
-            $props_valors = $request->input('propietats_valors');
 
             $propietat->save();
 
-            foreach ($var as $prop) {
-                $propBD = Propietat::where('nom', $prop)->first();
-                if ($propBD == null) {
-                    $propBD = new Propietat();
-                    $propBD->nom = $prop;
-                    $propBD->save();
+
+            if ($request->input('es_principal') == 1) {
+                $propietats = Propietats::where('id_article', $request->input('id_article'))->get();
+                foreach ($propietats as $p) {
+                    if ($p->id_propietats != $propietat->id_propietats) {
+                        $p->es_principal = 0;
+                        $p->save();
+                    }
                 }
-                $valor = Valors::where('valor', $props_valors[$prop])->first();
-                if ($valor == null) {
-                    $valor = new Valors();
-                    $valor->valor = $props_valors[$prop];
-                    $valor->save();
-                    $valor->propietat()->detach();
-                    $valor->propietat()->attach($propBD);
-                }
-                $propietat->valors()->attach($valor);
             }
+
+            if ($request->input('propietats_valors')) {
+                $var = array_keys($request->input('propietats_valors'));
+
+                $props_valors = $request->input('propietats_valors');
+
+                foreach ($var as $prop) {
+                    $propBD = Propietat::where('nom', $prop)->first();
+                    if ($propBD == null) {
+                        $propBD = new Propietat();
+                        $propBD->nom = $prop;
+                        $propBD->save();
+                    }
+
+                    foreach ($props_valors[$prop] as $prop) {
+                        $valor = Valors::where('valor', $prop)->first();
+                        if ($valor == null) {
+                            $valor = new Valors();
+                            $valor->valor = $prop;
+                            $valor->save();
+                            $valor->propietat()->detach();
+                            $valor->propietat()->attach($propBD);
+                        }
+                        $propietat->valors()->attach($valor);
+                    }
+                }
+            }
+
+
             $message = ($request->input('id_propietats') == null)
                 ? 'Propietat creada correctamente'
-                : 'Propietat actualizada correctamente';
-
-
+                : 'Propietat actualizada correctamente'; 
 
             return response()->json([
                 'message' => $message,
@@ -127,6 +145,7 @@ class PropietatsController extends Controller
             $imatge->save();
 
             $propietat = Propietats::find($id_prop);
+            $propietat->imatges()->detach();
             $propietat->imatges()->attach($imatge);
 
             $imatges = $request->file('imatgesSecundaries');
@@ -162,7 +181,7 @@ class PropietatsController extends Controller
         try {
             $propietat = Propietats::find($request->input('id_propietat'));
             $propietat->valors()->detach();
-            $propietat->imatges()->detach(); 
+            $propietat->imatges()->detach();
             $propietat->delete();
 
             return response()->json(['message' => 'Propietat eliminada correctamente'], 200);
